Enhanced parser process
To make the following processes easier, parsing process is changed. Now linked list implementation is being used, with a unified data structure interface for the all translator operations. To accommodate this, util functions are added in a separate file, with all the data structure related functions being in the data structure files.